requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
qrcode
reportlab
Werkzeug
email-validator
pytest
flake8
black
pika
